# docs/api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI 文章生成后端：用户/会话、文章三阶段流水线（SSE 进度）、配图、Agent 日志；开发态 VIP Mock 支付与管理端升降 VIP。",
        "title": "Wood Passage Creator API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/admin/article/list": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "全站文章列表，仅管理员；支持 pageNum/pageSize/status 查询参数。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-article"
                ],
                "summary": "分页查询全部文章（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认 1",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数，默认 10，最大 100",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "PENDING",
                            "PROCESSING",
                            "COMPLETED",
                            "FAILED"
                        ],
                        "type": "string",
                        "description": "状态筛选",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wood-passage-creator_internal_app_article.ArticleListData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
        "/article/confirm-outline": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "在 OUTLINE_EDITING 阶段确认大纲；仅文章作者或管理员可操作。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "确认文章大纲",
                "parameters": [
                    {
                        "description": "确认大纲参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_app_article.ConfirmOutlineRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功（data 一般为 null）",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误/阶段不允许",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "文章不存在",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/article/confirm-title": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "在 TITLE_SELECTING 阶段确认主/副标题；仅文章作者或管理员可操作。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "确认文章标题",
                "parameters": [
                    {
                        "description": "确认标题参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_app_article.ConfirmTitleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功（data 一般为 null）",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误/阶段不允许",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "文章不存在",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/article/create": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "登录用户创建一篇文章任务，进入生成流水线（具体异步逻辑由服务层处理）。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "创建文章生成任务",
                "parameters": [
                    {
                        "description": "创建参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_app_article.CreateArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wood-passage-creator_internal_module_article.Article"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/article/delete": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "按主键软删除；需登录。权限由业务层校验（作者或管理员）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "删除文章（软删除）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章主键 ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功（data 一般为 null）",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "文章不存在",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/article/execution-logs/{taskId}": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "返回该 task 下各 agent 步骤的耗时与状态汇总。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "任务执行日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务 ID",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wood-passage-creator_internal_module_article.AgentExecutionStats"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/article/list/self": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "仅返回当前登录用户自己的文章；支持 pageNum/pageSize/status 查询参数。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "分页查询我的文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认 1",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数，默认 10，最大 100",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "PENDING",
                            "PROCESSING",
                            "COMPLETED",
                            "FAILED"
                        ],
                        "type": "string",
                        "description": "状态筛选",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wood-passage-creator_internal_app_article.ArticleListData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/article/modify-outline": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "仅 VIP/管理员；阶段须为 OUTLINE_EDITING。按自然语言建议改大纲并落库。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "AI 修改大纲（VIP）",
                "parameters": [
                    {
                        "description": "任务 ID 与修改建议",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_app_article.AiModifyOutlineRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/wood-passage-creator_internal_module_article.OutlineSection"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "非 VIP 或阶段不允许",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/article/progress/{taskId}": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "按 taskId 建立 text/event-stream；需先登录且有权访问该任务。使用 SSE named events：connected / outline_delta / outline_done / task_error。",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "article"
                ],
                "summary": "SSE 订阅文章生成进度",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务 ID",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE 流",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "文章不存在",
                        "schema": {
                            "$ref": "#/definitions/wood-passage-creator_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "wood-passage-creator_internal_app_article.AiModifyOutlineRequest": {
            "type": "object",
            "required": [
                "modifySuggestion",
                "taskId"
            ],
            "properties": {
                "modifySuggestion": {
                    "type": "string",
                    "maxLength": 4000,
                    "minLength": 1
                },
                "taskId": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 1
                }
            }
        },
        "wood-passage-creator_internal_app_article.ArticleListData": {
            "type": "object",
            "properties": {
                "pageNum": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wood-passage-creator_internal_module_article.Article"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "wood-passage-creator_internal_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "SessionAuth": {
            "description": "Session cookie authentication. Example: session=your-session-id",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}